requirements.txt
# requirements.txt
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
pyyaml>=6.0